+++ database/migrations/2024_02_08_170013_update_messages_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         // Aggiornamento della tabella messages, aggiunta relazione tra apartment e message
         Schema::table('messages', function(Blueprint $table) {
-            $table->tinyInteger('apartment_id')->unsigned()->after('id');
+            $table->bigInteger('apartment_id')->unsigned()->after('id');
             $table->foreign('apartment_id')->references('id')->on('apartments')->cascadeOnDelete();
         });
     }
